refac: datadetailpage ~ing
