+++ test-web/src/components/Charts/choroplethMap/Helper/processMapData.js
@@ -58,3 +58,4 @@ export const processMapData = (data, mapData, setGeoJSONData, setKeyIdx) => {
   addProperties(cattleCnt, porkCnt, totalCnt);
   setKeyIdx((prev) => prev + 1);
 };
+export default processMapData;
